modules/locations/producer/requirements.txt
grpcio-tools==1.44.0
kafka-python==2.0.2
retry==0.9.2
